TKT-77: apply ai-review findings — slot precondition, expired-replay guard, loopback binds, int64 capacity math
Finding 1 (critical): ticket-type/slot agreement moves inside inventory's
locked conversion transaction as a precondition — a mismatch now rejects
with the operational hold untouched instead of after the carve committed.
Finding 2 (high, partial): commerce 409s a replay whose converted child
already expired rather than persisting a reservation no checkout can
complete; the full conversion-saga recommendation is rejected as out of
scope (the plan chose deterministic-ID replay; an expired child fails
checkout cleanly and its behavior is documented in ADR-023).
Finding 3 (high): staff ports bind to 127.0.0.1 only; removing the
repo-wide default internal token is pre-existing scope, deferred to a
backlog ticket.
Finding 4 (medium): capacity admission math widened to int64 in both
paths, with an int32-boundary oversell test.

+++ services/inventory/internal/store/store.go
@@ -176,7 +176,8 @@ func (p *Postgres) CreateHold(ctx context.Context, org, slot, ticketType uuid.UU
 	if err = tx.QueryRowContext(ctx, `SELECT COALESCE(sum(quantity),0) FROM claims WHERE pool_id=$1 AND `+liveClaims, slot).Scan(&held); err != nil {
 		return Claim{}, false, err
 	}
-	if confirmed+held+qty > capacity {
+	// int64 math: three valid int32 counters can wrap a 32-bit sum (ai-review finding 4).
+	if int64(confirmed)+int64(held)+int64(qty) > int64(capacity) {
 		return Claim{}, false, ErrUnavailable
 	}
 	c := Claim{ID: uuid.New(), OrganizerID: org, PoolID: slot, TicketTypeID: ticketType, Quantity: qty, UnitAmount: unitAmount, Currency: currency, Status: "held", Kind: "buyer"}
